+++ historian-server/src/main/java/com/hurence/webapiservice/http/api/grafana/parser/QueryRequestParser.java
@@ -35,13 +35,13 @@ public class QueryRequestParser {
         builder.withAdHocFilters(adHocFilters == null ? QueryRequestParam.DEFAULT_FILTERS : adHocFilters);
         String requestId = parseRequestId(requestBody);
         builder.withId(requestId);
-        Double quality = parseQuality(requestBody);
+        Float quality = parseQuality(requestBody);
         builder.withQuality(quality);
         return builder.build();
     }
 
-    private Double parseQuality(JsonObject requestBody) {
-        return requestBody.getDouble("quality");
+    private Float parseQuality(JsonObject requestBody) {
+        return requestBody.getFloat("quality");
     }
